Added SVM machine learning model for images classification
Added the SVM machine learning model needed for classifying the input from openCV images into 3 categories: Right (1), Left(2), Forward(0). The model was trained and tested on synthetic data generated from a python script uploaded too. The model accuracy was >0.99 with a high recall, percision and F1 score due to the fact of low to zero noise in the data and the dataset generated was not realistic.More model adjustments needed when obtaining the real data