Added Line following Model
Added prefinal model to detect lines with the specified outputs, Included all files used for cleaning the data, tuning openCV parameters, labelling, training the model and the Final Model. The final model file can be run by specifying a path to images and choosing any image and getting an output. The trained model is also loaded in the final model script along with the scalar used in training the model.